requirements.txt
numpy >= 1.26.0
absl-py >= 1.3.0
azure-cognitiveservices-vision-computervision >= 0.9.0
openai >= 0.25.0
requests >= 2.28.1
replicate >= 0.4.0
pillow >= 9.3.0
torch >= 1.13.0
diffusers >= 0.10.2
transformers >= 4.25.1
accelerate >= 0.15.0
pytesseract >= 0.3.10
deep-translator